% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{callr_profile}
\alias{callr_profile}
\title{callr-compatible function that monitors the process's resources}
\usage{
callr_profile(..., monitor_path, monitor_interval = 0.1)
}
\arguments{
\item{...}{Extra arguments are passed to the \link[processx:process]{processx::process}
constructor.}

\item{monitor_path}{Path to the RDS where the results will be saved. Character scalar.}

\item{monitor_interval}{Amount of time in seconds between resource polls. Double vector of length 1.}
}
\value{
The result of the function, identically to a standard callr function
}
\description{
callr-compatible function that monitors the process's resources
}
