% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{memory_plot}
\alias{memory_plot}
\title{Plot the memory usage over time}
\usage{
memory_plot(monitor_path, meta_args = list())
}
\arguments{
\item{monitor_path}{Path to the directory previously passed to [monitor_pid].
Character scalar.}
}
\value{
A `ggplot` object
}
\description{
Plot the memory usage over time
}
