% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{monitor_save}
\alias{monitor_save}
\title{Monitor a process and save the results}
\usage{
monitor_save(path, ...)
}
\arguments{
\item{path}{Path to the RDS where the results will be saved. Character scalar.}

\item{...}{
  Arguments passed on to \code{\link[=monitor_pid]{monitor_pid}}
  \describe{
    \item{\code{pid}}{Pid of the process to monitor. Integer vector with length 1.}
    \item{\code{interval}}{Amount of time in seconds between resource polls. Double vector of length 1.}
  }}
}
\value{
Nothing
}
\description{
Users will typically not need to call this function directly.
Instead, prefer \link{monitor_pid}.
}
