% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{process_meta}
\alias{process_meta}
\title{Add some time-based variables to targets metadata}
\usage{
process_meta(targets_metadata = targets::tar_meta())
}
\arguments{
\item{targets_metadata}{The result of \code{targets::tar_meta()} for your pipeline
of interest}
}
\value{
A copy of \code{targets_metadata}, with some additional columns
}
\description{
Add some time-based variables to targets metadata
}
\examples{
process_meta()
}
\concept{profiling-analysis}
