% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_data}
\alias{profile_data}
\title{Read the monitoring data}
\usage{
profile_data(monitor_path)
}
\arguments{
\item{monitor_path}{Path to the directory previously passed to \link{monitor_pid}.
Character scalar.}
}
\value{
Tibble containing monitoring data. Each row is one timepoint, and
the columns correspond to various process metrics.
}
\description{
Read the monitoring data
}
