% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{profile_per_target}
\alias{profile_per_target}
\title{Groups profile entries per target}
\usage{
profile_per_target(profile, targets = process_meta())
}
\arguments{
\item{profile}{A data frame obtained from \link{profile_data}}

\item{targets}{A data frame obtained from \link{process_meta}}
}
\value{
A data frame with all the columns from both \code{profile} and \code{target}
}
\description{
Groups profile entries per target
}
\concept{profiling-analysis}
