% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{summarise_targets}
\alias{summarise_targets}
\title{Calculate some resources summaries for each target}
\usage{
summarise_targets(combined)
}
\arguments{
\item{combined}{The result of \link{profile_per_target}}
}
\value{
A summary data frame, with one row per target. The columns include:
\itemize{
\item \code{name}: The target name
\item \code{peak_memory}: The maximum memory that this target used
\item \code{peak_memory_time}: The time at which the maximum memory usage was reached
\item \code{cpu_efficiency}: The percentage of the time that this target used the CPU
}
}
\description{
Calculate some resources summaries for each target
}
\concept{profiling-analysis}
