% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{monitor_save}
\alias{monitor_save}
\title{Monitor a process and save the results}
\usage{
monitor_save(path, ...)
}
\arguments{
\item{path}{Path to the RDS where the results will be saved. Character scalar.}
}
\value{
Nothing
}
\description{
Monitor a process and save the results
}
